add NS_PHY_INDIRECT flag in netmap_slot
It may be super vulnerable.
NS_PHY_INDIRECT is a flag for netmap_slot. When this flag is set 1, the
value of 'ptr' in netmap_slot is used as the "PHYSICAL" address of
the packet buffer. By using this option, we can use any addresses for
packet buffers. Currently I added it to only i40e netmap driver.